make Makefile.inc automatic, make it more generic
BSD Make has always automatically included any existing
"../Makefile.inc", so let it do so!

This partly reduces the need for ${bmake_topdir}, though not completely,
and it sorts out some other odd dependencies, e.g. now allowing it to
directly include Makefile.compiler.

An odd trick is needed to get it to work in two-or-more deep
sub-directories, which is also noticed on its inclusion of
Makefile.compiler:  notably ${.CURDIR} is not needed, and indeed
dangrou.  Hopefully this is not due to a bug in Make (though if it is
then it's easy to fix).

This change also tries to make Makefile.inc a wee bit more generic,
using LIB_* for version IDs instead of project-specific names.

And default targets get to move from Makefile.compiler back to
Makefile.inc again; and GENHDIR also moves to it.