deal with related CVE-2022-24795 from brianmario/yajl-ruby
The buffer expansion reallocation uses a binary left shift to increment
the size of the allocation.  This will of course overflow (by "wrapping"
to zero) when the buffer size reaches 2^((sizeof(size_t)*CHAR_BIT)-1).

Since this library doesn't do its own memory allocation error checking
the only current fix is to add an assert() check and to recommend that
user-supplied allocation routines also check for realloc() requests with
a size of zero and if so treat that as an error as well.