doc/Makefile: wrap cxref script in a subshell
It is good practice to wrap commands that change the current directory
in a sub-shell so that they don't affect any subsequent commands as BSD
Makes handle multiple commands differently depending on whether they are
in "jobs" mode or not.

This doesn't actually matter here because the subsequent command creates
a link with a fully qualified pathname so it won't matter where it is
run from.